Refactored config to comply with new conventions specified in https://dans-knaw.github.io/dans-module-archetype/dans-module-design/\#java-packages
