+++ src/main/java/nl/knaw/dans/sword2/config/AuthorizationConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.core.config;
+package nl.knaw.dans.sword2.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
