+++ src/main/java/nl/knaw/dans/sword2/config/CollectionConfig.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.core.config;
+package nl.knaw.dans.sword2.config;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import nl.knaw.dans.sword2.core.DepositState;
-import nl.knaw.dans.sword2.core.config.converter.StringByteSizeConverter;
+import nl.knaw.dans.sword2.config.converter.StringByteSizeConverter;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
