+++ src/main/java/nl/knaw/dans/sword2/config/SwordError.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package nl.knaw.dans.sword2.core.config;
+package nl.knaw.dans.sword2.config;
 
 import javax.xml.bind.annotation.XmlEnum;
 import javax.xml.bind.annotation.XmlEnumValue;
