+++ src/main/java/nl/knaw/dans/sword2/config/UserConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.core.config;
+package nl.knaw.dans.sword2.config;
 
 import javax.validation.constraints.NotEmpty;
 import java.util.List;
