+++ src/main/java/nl/knaw/dans/sword2/config/converter/StringByteSizeConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.core.config.converter;
+package nl.knaw.dans.sword2.config.converter;
 
 import com.fasterxml.jackson.databind.util.StdConverter;
