+++ src/main/java/nl/knaw/dans/sword2/core/service/CollectionManager.java
@@ -16,7 +16,7 @@
 package nl.knaw.dans.sword2.core.service;
 
 import nl.knaw.dans.sword2.core.auth.Depositor;
-import nl.knaw.dans.sword2.core.config.CollectionConfig;
+import nl.knaw.dans.sword2.config.CollectionConfig;
 import nl.knaw.dans.sword2.core.exceptions.CollectionNotFoundException;
 
 import java.util.List;
