+++ src/main/java/nl/knaw/dans/sword2/core/service/DepositHandlerImpl.java
@@ -18,8 +18,8 @@ package nl.knaw.dans.sword2.core.service;
 import nl.knaw.dans.sword2.core.Deposit;
 import nl.knaw.dans.sword2.core.DepositState;
 import nl.knaw.dans.sword2.core.auth.Depositor;
-import nl.knaw.dans.sword2.core.config.CollectionConfig;
-import nl.knaw.dans.sword2.core.config.UriRegistry;
+import nl.knaw.dans.sword2.config.CollectionConfig;
+import nl.knaw.dans.sword2.config.UriRegistry;
 import nl.knaw.dans.sword2.core.exceptions.CollectionNotFoundException;
 import nl.knaw.dans.sword2.core.exceptions.DepositNotFoundException;
 import nl.knaw.dans.sword2.core.exceptions.DepositReadOnlyException;
