+++ src/main/java/nl/knaw/dans/sword2/core/service/UserManagerImpl.java
@@ -16,7 +16,7 @@
 package nl.knaw.dans.sword2.core.service;
 
 import nl.knaw.dans.sword2.core.auth.Depositor;
-import nl.knaw.dans.sword2.core.config.UserConfig;
+import nl.knaw.dans.sword2.config.UserConfig;
 
 import java.util.HashMap;
 import java.util.List;
