+++ src/test/java/nl/knaw/dans/sword2/core/service/DepositHandlerImplTest.java
@@ -19,8 +19,8 @@ import nl.knaw.dans.sword2.TestFixture;
 import nl.knaw.dans.sword2.core.Deposit;
 import nl.knaw.dans.sword2.core.DepositState;
 import nl.knaw.dans.sword2.core.auth.Depositor;
-import nl.knaw.dans.sword2.core.config.CollectionConfig;
-import nl.knaw.dans.sword2.core.config.Sword2Config;
+import nl.knaw.dans.sword2.config.CollectionConfig;
+import nl.knaw.dans.sword2.config.Sword2Config;
 import nl.knaw.dans.sword2.core.exceptions.CollectionNotFoundException;
 import nl.knaw.dans.sword2.core.exceptions.DepositNotFoundException;
 import nl.knaw.dans.sword2.core.exceptions.InvalidDepositException;
@@ -29,13 +29,11 @@ import nl.knaw.dans.sword2.core.exceptions.NotEnoughDiskSpaceException;
 import nl.knaw.dans.sword2.core.finalizer.DepositFinalizerEvent;
 import org.apache.commons.io.FileUtils;
 import org.eclipse.jetty.util.BlockingArrayQueue;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.time.OffsetDateTime;
