+++ src/test/java/nl/knaw/dans/sword2/resources/CollectionResourceBagSenderIntegrationTest.java
@@ -16,14 +16,9 @@
 package nl.knaw.dans.sword2.resources;
 
 import ch.qos.logback.classic.LoggerContext;
-import io.dropwizard.testing.ResourceHelpers;
-import io.dropwizard.testing.junit5.DropwizardAppExtension;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import net.lingala.zip4j.ZipFile;
 import net.lingala.zip4j.model.ZipParameters;
-import nl.knaw.dans.sword2.DdSword2Application;
-import nl.knaw.dans.sword2.DdSword2Configuration;
-import nl.knaw.dans.sword2.TestFixture;
 import nl.knaw.dans.sword2.TestFixtureExt;
 import nl.knaw.dans.sword2.api.entry.Entry;
 import nl.knaw.dans.sword2.core.Deposit;
@@ -52,7 +47,6 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.security.NoSuchAlgorithmException;
 
 @ExtendWith(DropwizardExtensionsSupport.class)
