+++ src/test/java/nl/knaw/dans/sword2/resources/CollectionResourceLimitedDiskSpaceIntegrationTest.java
@@ -16,15 +16,9 @@
 package nl.knaw.dans.sword2.resources;
 
 import ch.qos.logback.classic.LoggerContext;
-import io.dropwizard.testing.ResourceHelpers;
-import io.dropwizard.testing.junit5.DropwizardAppExtension;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
-import nl.knaw.dans.sword2.DdSword2Application;
-import nl.knaw.dans.sword2.DdSword2Configuration;
-import nl.knaw.dans.sword2.TestFixture;
 import nl.knaw.dans.sword2.TestFixtureExt;
 import nl.knaw.dans.sword2.core.service.FileServiceImpl;
-import org.apache.commons.io.FileUtils;
 import org.glassfish.jersey.media.multipart.MultiPartFeature;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -35,8 +29,6 @@ import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.Invocation;
 import javax.ws.rs.core.MediaType;
 import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
