+++ src/test/java/nl/knaw/dans/sword2/resources/StatementResourceImplIntegrationTest.java
@@ -16,14 +16,7 @@
 package nl.knaw.dans.sword2.resources;
 
 import ch.qos.logback.classic.LoggerContext;
-import io.dropwizard.configuration.FileConfigurationSourceProvider;
-import io.dropwizard.configuration.SubstitutingSourceProvider;
-import io.dropwizard.testing.ResourceHelpers;
-import io.dropwizard.testing.junit5.DropwizardAppExtension;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
-import nl.knaw.dans.sword2.DdSword2Application;
-import nl.knaw.dans.sword2.DdSword2Configuration;
-import nl.knaw.dans.sword2.TestFixture;
 import nl.knaw.dans.sword2.TestFixtureExt;
 import nl.knaw.dans.sword2.api.statement.Feed;
 import nl.knaw.dans.sword2.core.Deposit;
@@ -32,7 +25,6 @@ import nl.knaw.dans.sword2.core.exceptions.InvalidDepositException;
 import nl.knaw.dans.sword2.core.service.DepositPropertiesManagerImpl;
 import nl.knaw.dans.sword2.core.service.FileServiceImpl;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.text.StringSubstitutor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -41,7 +33,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import java.io.IOException;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
-import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
